Fix the issue that disable WiFi tethering leads to callbacks lost
Root cause:
When WiFi tethering is disabled, WiFi technology will be removed and
added again in ConnMan layer, which leads to the callbacks saved in
ConnMan lib layer are released, so that it will never invoke to call
the CAPI registered callbacks.

Solution:
In CAPI layer, when it detects that technology is added, set the
WiFi tethering callbacks to ConnMan lib layer again if necessary.

Change-Id: I902d0d982c7751e367c3e26d50f5155e65534b38